refactor: extract benefit statement comparators into API
